Migrate get_sex to separate file
